+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog\
 
 ## [Unreleased]
+* Hotfix - Remove ID word on label for the channel input text to prevent autocomplete from browsers.
 * Feature - discordReactionManager: user from the message in the output, complete flow message is being copied to the output
 * Feature - Ephemeral check for commands (Auto reply)
 * Hotfix - Null validation when trying to access to msg.payload properties